% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideogram.R
\name{ideogram}
\alias{ideogram}
\title{R binding for ideogram}
\usage{
ideogram(file, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{file}{IBIS IBD segments file}
}
\description{
Plots ideogram.
}
