% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2ggplot.R
\name{pdf2ggplot}
\alias{pdf2ggplot}
\title{pdf2ggplot}
\usage{
pdf2ggplot(pdf, cut_empty_space = TRUE, density = 600)
}
\arguments{
\item{pdf}{pdf file}

\item{cut_empty_space}{whether cut empty margins}

\item{density}{density to parse the pdf file}
}
\value{
ggplot object
}
\description{
read pdf and draw the content using ggplot
}
\author{
Guangchuang Yu
}
