% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideogram.R
\name{readFile}
\alias{readFile}
\title{Reads input IBD segment file}
\usage{
readFile(file)
}
\arguments{
\item{file}{input file}
}
\value{
A data frame containing IBD segment data from the input file
}
\description{
Reads input IBD segment file
}
