% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg2pdf.R
\name{svg2pdf}
\alias{svg2pdf}
\title{svg2pdf}
\usage{
svg2pdf(svg, pdf)
}
\arguments{
\item{svg}{svg file}

\item{pdf}{pdf file}
}
\value{
invisible grob object
}
\description{
convert svg to pdf
}
\author{
Guangchuang Yu
}
