% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideogRam.R
\name{ideogRam-shiny}
\alias{ideogRam-shiny}
\alias{ideogRamOutput}
\alias{renderIdeogRam}
\title{Shiny bindings for ideogRam}
\usage{
ideogRamOutput(outputId, width = "100\%", height = "400px")

renderIdeogRam(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a ideogRam}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using ideogRam within Shiny
applications and interactive Rmd documents.
}
\examples{
server <- function(input, output) {
    get_config <- reactive({
        list(
            organism = "human",
            orientation = "horizontal",
            chromosome = as.character(input$chr),   ## here is the issue....
            annotations = list(
                list(
                    name = "xxx",
                    chr = "4",
                    start = 34294,
                    stop = 125482
                ),
                list(
                    name = "BRCA1",
                    chr = "17",
                    start = 43125400,
                    stop = 43125482
                )
            ),
            # Note: for newer version of ideogram, it should be a character like "chr4:34294-1125482"
            brush = TRUE
        )
    })
    output$ideo_01 <- renderIdeogRam({
        ideo_01_config <- get_config()
        p <- ideogRam(data = ideo_01_config) # elementId has to be the same with output$ obj name
        p
    })
    output$region_desc <- renderText({
        rg <- input$ideo_01_brushrange
        paste0("From: ", rg$from, ";\\n", "To: ", rg$to, ";\\n", "Extent: ", rg$extent, ";\\n")
    })
}

ui <- shinyUI(fluidPage(
    titlePanel("Hello IdeogRam!"),
    sidebarPanel(
        width = 2,
        numericInput("chr", "Chromosome:", 4, min = 1, max = 26)
    ),
    mainPanel(
        ideogRamOutput("ideo_01"),
        verbatimTextOutput("region_desc")
    )
))
runApp(list(ui = ui, server = server))
}
