% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideogram.R
\name{ideogram}
\alias{ideogram}
\title{ideogram}
\usage{
ideogram(karyotype, overlaid, label = NULL, output = "chromosome.svg")
}
\arguments{
\item{karyotype}{karyotype data}

\item{overlaid}{overlaid data}

\item{label}{track label data}

\item{output}{output file, only svg is supported}
}
\value{
output file
}
\description{
ideogram with overlaid heatmap annotation and optional track label
}
\author{
Zhaodong Hao, Dekang Lv and Guangchuang Yu
}
